Add more pedantic warnings, and fix code to match them

+++ chartype.h
@@ -21,14 +21,14 @@
 #define C_TERM 0x0040   // terminates an identifier
 #define C_INCL 0x0080   // terminates ident, but included
 
-#define IsSep(c) (cType[c] & C_SEP)
-#define IsTok(c) (cType[c] & C_TOK)
-#define IsDigit(c) (cType[c] & C_DIGIT)
-#define IsBin(c) (cType[c] & C_BIN)
-#define IsDec(c) (cType[c] & C_DEC)
-#define IsHex(c) (cType[c] & C_HEX)
-#define IsTerm(c) (cType[c] & C_TERM)
-#define IsIncl(c) (cType[c] & C_INCL)
+#define IsSep(c) (cType[static_cast<uint8_t>(c)] & C_SEP)
+#define IsTok(c) (cType[static_cast<uint8_t>(c)] & C_TOK)
+#define IsDigit(c) (cType[static_cast<uint8_t>(c)] & C_DIGIT)
+#define IsBin(c) (cType[static_cast<uint8_t>(c)] & C_BIN)
+#define IsDec(c) (cType[static_cast<uint8_t>(c)] & C_DEC)
+#define IsHex(c) (cType[static_cast<uint8_t>(c)] & C_HEX)
+#define IsTerm(c) (cType[static_cast<uint8_t>(c)] & C_TERM)
+#define IsIncl(c) (cType[static_cast<uint8_t>(c)] & C_INCL)
 
 #define _tolower(c) ((c) | ('a' - 'A'))
 #define islower(c) ('a' <= (c) && (c) <= 'z')
